genLogRegData {prob}	R Documentation

Generate data for logistic regression

Description

This function generates data ready for a logistic regression model

Usage

genLogRegData(xdata, beta = rep(1, ncol(xdata)), yname = "y")

Arguments

xdata
the model matrix

beta
vector of parameters to multiply the model matrix

yname
the name for the generated y values

Details

This function generates data ready for a logistic regression model

Value

A data frame with the model matrix and the generated y values added

Author(s)

G. Jay Kerns

[Package prob version 1.0-1 Index]