genXdata {prob}	R Documentation

Generate continuous model matrix data

Description

This function generates correlated normal data to serve as a model matrix in a regression model.

Usage

genXdata(n, nvar = 1, mu = rep(0, nvar), Sigma = diag(length(mu)), 
varnames = paste("x", 1:length(mu), sep = ""), roundto = NULL)

Arguments

n
how many rows

nvar
how many columns

mu
the mean of the multivariate normal distribution

Sigma
the variance-covariance matrix of the normal distribution

varnames
how you would like the variables to be named in the result

roundto
number of places to round the generated values

Details

This function generates correlated normal data to serve as a model matrix in a regression model.

Value

A data frame of generated data

Author(s)

G. Jay Kerns

[Package prob version 1.0-1 Index]