probspace {prob}	R Documentation

Probability Spaces

Description

Forms a probability space from a set of outcomes and (optional) vector of probabilities.

Usage

probspace(x, ...)

## Default S3 method:
probspace(x, probs, ...)

## S3 method for class 'list'
probspace(x, probs, ...)

Arguments

x
a vector, data frame, or list of outcomes.

probs
a vector of non-negative weights of the same length as outcomes

...
further arguments to be passed to or from other methods.

Details

The elements of probs will be normalized to ensure that their sum is one. If probs is not specified, then the equally likely model is assumed in which every outcome has the same probability.

Value

If outcomes is a vector or data frame, then the value is a data frame with an added probs column. If outcomes is a list, then the value is a list with components outcomes (the supplied list) and a probs component.

Author(s)

G. Jay Kerns gkerns@ysu.edu.

See Also

iidspace

Examples

R <- rolldie(3)
probspace(R)

[Package prob version 1.0-1 Index]