union {prob}	R Documentation

Union of Subsets

Description

Calculates the union of subsets of a probability space.

Usage

union(x, ...)

## Default S3 method:
union(x, y, ...)

## S3 method for class 'data.frame'
union(x, y, ...)

## S3 method for class 'ps'
union(x, y, ...)

Arguments

x, y
vectors, data frames, or ps objects containing a sequence of items (conceptually)

...
further arguments to be passed to or from other methods.

Details

This function operates row-wise on dataframes, and element-wise among the outcomes of ps objects. The elements of union(x,y) are those elements in x or y, or both. The definition is taken to match the version in the base package.

Value

A data frame or subset of a probability space of the same type as its arguments.

Author(s)

G. Jay Kerns gkerns@ysu.edu, based on a suggestion made by Brian Ripley in R-devel, 12/11/07.

See Also

intersect, setdiff

Examples

S <- cards()
A <- subset(S, suit == "Heart")
B <- subset(S, rank == "A" )
union(A, B)
   

[Package prob version 1.0-1 Index]