cards(prob)	R Documentation

A Standard Set of Playing Cards

Description

The title says it all.

Usage


cards(jokers = FALSE, makespace = FALSE)


Arguments

jokers 	 logical.

makespace 	 logical.

Details

This generates a data frame sample space of a standard deck of 52 playing cards. Optionally, the user can specify that Jokers be included, which have a rank but with suit a missing value.

Value

A data frame with columns rank and suit, and optionally a column of equally likely probs.

See Also

rolldie, tosscoin, and roulette

Examples


cards()
cards(makespace = TRUE)

[Package prob version 0.9-1 Index]