euchredeck(prob)	R Documentation

A Deck of Playing Cards for Euchre

Description

The title says it all.

Usage


euchredeck(benny= FALSE, makespace = FALSE)


Arguments

benny 	 logical.

makespace 	 logical.

Details

This is a conventional Euchre deck which uses a deck of 24 standard playing cards consisting of Ace, King, Queen, Jack, 10, and 9 of each of the four suits. If benny = TRUE then a Joker is added to the deck.

Value

A data frame with columns value and suit, and optionally a column of equally likely probs.

See Also

cards, tosscoin, and roulette

Examples


euchredeck()
euchredeck(benny = TRUE, makespace = TRUE)

[Package prob version 0.9-1 Index]