nsamp(prob)	R Documentation

Number of Samples from an Urn

Description

Calculates the number of samples from an urn under different sampling scenarios.

Usage


nsamp(n, k, replace = FALSE, ordered = FALSE)


Arguments

n 	 an integer or integer vector.

k 	 an integer or integer vector.

replace 	 logical indicating whether sampling should be done with replacement.

ordered 	 logical indicating whether order among samples is important.

Details

The nsamp() function will calculate the number of samples from an urn under assorted assumptions on the sampling procedure. The arguments are: n, the number of (distinguishable) objects in the urn, k, the sample size, and replace, ordered as documented in urnsamples.

nsamp() is vectorized, so that entering vectors instead of numbers for n, k, replace, and ordered results in a vector of corresponding answers.

The formulas used in the four possible combinations of replace and ordered are as follows.

When replace = TRUE and ordered = TRUE, the value is n^k.
When replace = FALSE and ordered = TRUE, the value is n!/(n-k)!.
When replace = FALSE and ordered = FALSE, the value is n!/[k!(n-k)!].
When replace = TRUE and ordered = FALSE, the value is (n-1+k)!/[(n-1)!k!].

Value

A number.

Author(s)

G. Jay Kerns gkerns@ysu.edu.

See Also

urnsamples

Examples


nsamp(n=3, k=2, replace = TRUE, ordered = TRUE)
nsamp(n=3, k=2, replace = TRUE, ordered = FALSE)
nsamp(n=3, k=2, replace = FALSE, ordered = FALSE)
nsamp(n=3, k=2, replace = FALSE, ordered = TRUE)

[Package prob version 0.9-1 Index]