permsn(prob)	R Documentation

Generate All Permutations of x Elements Taken m at a Time

Description

Generate all permutations of the elements of x taken m at a time. If x is a positive integer, returns all permutations of the elements of seq(x) taken m at a time.

Usage


permsn(x, m)

Arguments

x 	 vector source for permutations, or integer n for x <- seq(n).

m 	 number of elements to permute.

Value

a list or array (in nondegenerate cases).

Author(s)

G. Jay Kerns gkerns@ysu.edu, modified from the combn function in the package utils.

See Also

combn

Examples


permsn(3,2)

[Package prob version 0.9-1 Index]