roulette(prob)	R Documentation

Roulette

Description

Sets up a sample space for the experiment of spinning a roulette wheel once.

Usage


roulette(european = FALSE, makespace = FALSE)

Arguments

european 	 logical.

makespace 	 logical.

Details

If european is TRUE, then a traditional EU roulette wheel with 37 pockets is used, otherwise, a standard US roulette wheel with 38 pockets is used. Entries in the data frame are ordered in the customary way to facilitate the calculation probabilities regarding called bets.

Value

A data frame, with columns num and color, and an equally likely probs column if makespace is TRUE.

Author(s)

G. Jay Kerns gkerns@ysu.edu.

See Also

cards

Examples


roulette()
roulette(european = TRUE, makespace = TRUE)

[Package prob version 1.0-1 Index]