subset(prob)	R Documentation

Subsets of Probability Spaces

Description

This is a method for subset() for the case when the input object is a probability space of class ps.

Usage


subset(x, ...)

## S3 method for class 'ps':
subset(x, subset, ...)


Arguments

x 	 a probability space.

subset 	 logical expression indicating elements or rows of space to keep: missing values are taken as false.

... 	 further arguments to be passed to or from other methods.

Details

This function simply extends the existing subset() function to ps objects.

Value

A ps object, a subset of a probability space.

Author(s)

G. Jay Kerns gkerns@ysu.edu.

See Also

intersect, setdiff, union, isin

Examples


L <- tosscoin(2)
M <- urnsamples(L, 3)
N <- probspace(M)
subset(N, all(toss1=="H"))
subset(N, any(toss2=="T"))

[Package prob version 0.9-1 Index]